requirements.txt
certifi==2019.9.11
chardet==3.0.4
cloudinary==1.18.2
DateTime==4.3
defusedxml==0.6.0
Django==2.2.6
django-admin==2.0.0
django-braces==1.13.0
django-cors-headers==3.1.1
django-excel-response2==2.0.8
django-oauth-toolkit==1.2.0
django-rest-framework==0.1.0
django-rest-framework-social-oauth2==1.1.0
django-six==1.0.4
django-widget-tweaks==1.4.5
djangorestframework==3.9.3
djangorestframework-expiring-authtoken==0.1.4
gunicorn==19.9.0
idna==2.8
mock==3.0.5
oauthlib==3.1.0
psycopg2==2.8.3
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.2.0
screen==1.0.1
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
urllib3==1.25.6
whitenoise==4.1.4
xlwt==1.3.0
zope.interface==4.6.0
